Refactor OptimizeIndexing to make it more extensible
This PR decouples the logic necessary to compute bounds on variables
from the logic that uses this info to perform the strenght analysis on
int64 variables. While doing so, it tries to minimize the number of
attributes of the class in favour of local variables.

This class is now accessible from any `LoopBody` object.

ghstack-source-id: b96242eade49eb3d55c77d5881e832f8ce35f763
Pull Request resolved: https://github.com/pytorch/pytorch/pull/100549